apirest/tests/requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asyncmy==0.2.10
asyncua==1.1.8
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
fastapi==0.117.1
freeopcua==0.90.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lxml==6.0.2
mysqlclient==2.2.7
opcua==0.98.13
packaging==25.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
